+++ modules/sde-external-services/edc/src/main/java/org/eclipse/tractusx/sde/edc/util/EDCAssetLookUp.java
@@ -86,5 +86,4 @@ public class EDCAssetLookUp {
 
 		return offers;
 	}
-
 }