+++ modules/sde-submodules/assembly-part-relationship/src/main/java/org/eclipse/tractusx/sde/submodels/apr/steps/EDCAspectRelationshipHandlerUseCase.java
@@ -62,7 +62,7 @@ public class EDCAspectRelationshipHandlerUseCase extends Step {
 		try {
 
 			AssetEntryRequest assetEntryRequest = assetFactory.getAssetRequest(submodel,
-					getSubmodelShortDescriptionOfModel(), shellId, subModelId, input.getParentUuid());
+					getSubmodelShortDescriptionOfModel(), shellId, subModelId, input.getParentUuid(), null, null, null);
 
 			Map<String, String> eDCAsset = null;