+++ modules/sde-submodules/single-level-usage-as-built/src/main/java/org/eclipse/tractusx/sde/submodels/sluab/steps/EDCSingleLevelUsageAsBuiltHandlerUseCase.java
@@ -61,7 +61,7 @@ public class EDCSingleLevelUsageAsBuiltHandlerUseCase extends Step {
 		try {
 
 			AssetEntryRequest assetEntryRequest = assetFactory.getAssetRequest(submodel,
-					getSubmodelShortDescriptionOfModel(), shellId, subModelId, input.getParentUuid());
+					getSubmodelShortDescriptionOfModel(), shellId, subModelId, input.getParentUuid(), null, null, null);
 
 			Map<String, String> eDCAsset =null;