+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Support for pcf v6.0.0 submodel.
 - Added controller interface api's for PCF Exchange.
 - Added new classes for multiple submodel version support.
+- Added refactor code changes for external services.
 - Added different usecase Handler for multi version support.
 
 ## [2.3.6] - 2024-03-06